refactor: relocate SDLC framework code to framework src/ directory
Restructure codebase to separate global tooling from framework-specific code:

MOVED to agentic/code/frameworks/sdlc-complete/src/:
- analysis/ - Codebase analyzer (UC-003)
- traceability/ - Requirements traceability (UC-006)
- security/ - Security validation (UC-011)
- orchestration/ - Multi-agent SDLC orchestration
- git/ - Git workflow orchestration
- cicd/ - CI/CD pipeline generation
- metrics/ - DORA metrics, project tracking
- monitoring/ - Performance monitoring
- recovery/ - Error recovery
- testing/ - NFR test infrastructure, mocks, fixtures

REMAIN in root src/ (global):
- cli/ - Global CLI orchestration
- agents/ - Agent deployment across frameworks
- plugin/ - Framework/workspace management
- writing/ - Writing Guide framework code

Also:
- Added framework-specific tsconfig.json
- Updated root tsconfig.json with path aliases (@sdlc/*, @global/*)
- Moved SDLC tests to test/unit/sdlc/
- Updated all import paths
- Updated framework README.md with new src/ documentation
