+++ agentic/code/frameworks/sdlc-complete/src/analysis/codebase-analyzer.ts
