+++ agentic/code/frameworks/sdlc-complete/src/recovery/error-recovery.ts
