+++ agentic/code/frameworks/sdlc-complete/src/security/api-patterns.ts
