+++ agentic/code/frameworks/sdlc-complete/src/security/secret-patterns.ts
