+++ agentic/code/frameworks/sdlc-complete/src/security/security-validator.ts
