+++ agentic/code/frameworks/sdlc-complete/src/testing/mocks/filesystem-sandbox.ts
