+++ agentic/code/frameworks/sdlc-complete/src/testing/mocks/git-sandbox.ts
