+++ agentic/code/frameworks/sdlc-complete/src/testing/mocks/github-stub.ts
