+++ agentic/code/frameworks/sdlc-complete/src/testing/trend-analyzer.ts
