+++ agentic/code/frameworks/sdlc-complete/src/traceability/traceability-checker.ts
