+++ test/unit/plugin/workspace-creator.test.ts
@@ -9,7 +9,7 @@
 
 import { describe, it, expect, beforeEach, afterEach } from 'vitest';
 import { WorkspaceCreator } from '../../../src/plugin/workspace-creator.js';
-import { FilesystemSandbox } from '../../../src/testing/mocks/filesystem-sandbox.js';
+import { FilesystemSandbox } from '../../../agentic/code/frameworks/sdlc-complete/src/testing/mocks/filesystem-sandbox.js';
 
 describe('WorkspaceCreator', () => {
   let sandbox: FilesystemSandbox;
