+++ test/unit/sdlc/testing/mocks/agent-orchestrator.test.ts
@@ -10,7 +10,7 @@ import {
   type MockAgentBehavior,
   type AgentRequest,
   type AgentResponse
-} from '../../../../src/testing/mocks/agent-orchestrator.js';
+} from '../../../../../../agentic/code/frameworks/sdlc-complete/src/testing/mocks/agent-orchestrator.js';
 
 describe('MockAgentOrchestrator', () => {
